requirements.txt
crewai==0.79.0
crewai-tools==0.14.0
flask==3.0.3
flask-cors==5.0.0
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-search-results==2.1.0
Jinja2==3.1.4
kaleido==0.2.1
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.51
langchain-experimental==0.3.3
langchain-openai==0.2.6
langchain_groq==0.3.2
langgraph==0.2.45
langserve==0.3.0
langsmith==0.1.142
numpy==1.26.4
openai==1.55.3
pandas==2.2.3
pydantic==2.9.2
PyPDF2==3.0.1
python-box==7.2.0
python-dotenv==1.0.1
requests==2.32.3
sendgrid==6.11.0
tavily-python==0.5.0
langchain-ollama==0.2.0